Update LED_Button.c
LED was dim because in the while loop, the LED was set to HIGH and was immediately turning it off if the button was not pressed.

This version makes the LED stay on as long as the button is pressed, and it will turn off when the button is released. This ensures that the LED is fully illuminated when the button is held down.